pipelines/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
supabase
pandas==2.2.0
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
geopy==2.4.1
